+++ examples/mean/app.py
@@ -1,9 +1,9 @@
 import statistics
 
-import pycape
+import serdio
 
 
-@pycape.lift_io
+@serdio.lift_io
 def compute_mean(x):
     return statistics.mean(x)
 
@@ -11,4 +11,4 @@ def compute_mean(x):
 cape_handler = compute_mean.as_cape_handler()
 
 # NOTE: this would have also worked, since `compute_mean` is trivial
-#  cape_handler = pycape.lift_io(statistics.mean).as_cape_handler()
+#  cape_handler = serdio.lift_io(statistics.mean).as_cape_handler()
