--- lift_io/lift_io/__init__.py
@@ -1,13 +1,0 @@
-from io_lifter import lift_io
-from serde import SerdeHookBundle
-from serde import bundle_serde_hooks
-from serde import deserialize
-from serde import serialize
-
-__all__ = [
-    "lift_io",
-    "bundle_serde_hooks",
-    "deserialize",
-    "serialize",
-    "SerdeHookBundle",
-]
