+++ serdio/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "lift_io"
+name = "serdio"
 description = "Auto-magic lifting of function inputs + outputs into msgpack-ed bytes."
 version = "0.1.0"
 requires-python = ">=3.7"
